Add some more tests

+++ test/test.sh
@@ -7,6 +7,13 @@ function call {
   body_line_count=`echo -e "$body" | wc --line`
 }
 
+function call_with_data {
+  local output=`curl --request $1 --data "$2" --silent --write-out "\n%{http_code}\n" $3`
+  result=`echo -e "$output" | tail -n 1`
+  body=`echo -e "$output" | head -n -1`
+  body_line_count=`echo -e "$body" | wc --line`
+}
+
 # Recreate the environment
 rm -rf sandbox*
 rm -f server*.log
@@ -16,8 +23,14 @@ mkdir sandbox2
 # Start the servers
 server1="127.0.0.1:2020"
 server2="127.0.0.1:2021"
-pushd sandbox1 && ../../httpfs -p 2020 $server2 > ../out1.log &
-pushd sandbox2 && ../../httpfs -p 2021 $server1 > ../out2.log &
+cd sandbox1
+../../httpfs -p 2020 $server2 > ../out1.log &
+server1_pid=$!
+cd ..
+cd sandbox2
+../../httpfs -p 2021 $server1 > ../out2.log &
+server2_pid=$!
+cd ..
 
 sleep 1
 
@@ -113,4 +126,65 @@ then
   exit 1
 fi
 
-killall httpfs
+echo "Check writing to file works"
+
+some_text="Lorem ipsum"
+
+cmd="call_with_data PUT '$some_text' $server1/foo"
+eval $cmd
+if [ "$result" -ne 200 ];
+then
+  echo "$cmd"
+  echo -e "$result"
+  echo -e "$body"
+  killall httpfs
+  exit 1
+fi
+
+cmd="call GET $server1/foo"
+eval $cmd
+if [ "$result" -ne 200 ] || [ "$body" != "$some_text" ];
+then
+  echo $cmd
+  echo -e "$result"
+  echo -e "$body"
+  killall httpfs
+  exit 1
+fi
+
+cmd="call GET $server2/foo"
+eval $cmd
+if [ "$result" -ne 200 ] || [ "$body" != "$some_text" ];
+then
+  echo $cmd
+  echo -e "$result"
+  echo -e "$body"
+  killall httpfs
+  exit 1
+fi
+
+echo "Check deleting file works"
+
+cmd="call DELETE $server1/foo"
+eval $cmd
+if [ "$result" -ne 200 ];
+then
+  echo "$cmd"
+  echo -e "$result"
+  echo -e "$body"
+  killall httpfs
+  exit 1
+fi
+
+cmd="call GET $server1/foo"
+eval $cmd
+if [ "$result" -ne 404 ]
+then
+  echo $cmd
+  echo -e "$result"
+  echo -e "$body"
+  killall httpfs
+  exit 1
+fi
+
+kill -9 $server1_pid $server2_pid
